method issueReportUseCase now working; method createInventoryAsset changed to InMemoryInventoryAsset
