+++ src/main/java/br/edu/ifsp/addthenewsoul/application/repository/imMemory/InMemoryAssetDAO.java
@@ -44,17 +44,6 @@ public class InMemoryAssetDAO implements AssetDAO {
     }
 
     @Override
-    public List<InventoryAsset> createInventoryAsset(List<Asset> assets) {
-        List<InventoryAsset> inventoryAssets = new ArrayList<>();
-        for (Asset asset : assets) {
-            inventoryAssets.add(new InventoryAsset(asset.getId(), asset.getDescription(),
-                    asset.getEmployeeInCharge(), asset.getValue(), asset.getDamage(), asset.getLocation(),
-                    Status.NOT_VERIFIED));
-        }
-        return inventoryAssets;
-    }
-
-    @Override
     public Integer add(Asset asset) {
         dbMemoryAsset.put(asset.getId(), asset);
         return asset.getId();
