+++ src/main/java/br/edu/ifsp/addthenewsoul/application/repository/imMemory/InMemoryInventoryDAO.java
@@ -1,5 +1,6 @@
 package br.edu.ifsp.addthenewsoul.application.repository.imMemory;
 
+import br.edu.ifsp.addthenewsoul.domain.entities.asset.Asset;
 import br.edu.ifsp.addthenewsoul.domain.entities.employee.Employee;
 import br.edu.ifsp.addthenewsoul.domain.entities.employee.Role;
 import br.edu.ifsp.addthenewsoul.domain.entities.inventory.Inventory;
@@ -82,4 +83,14 @@ public class InMemoryInventoryDAO implements InventoryDAO {
     }
 
 
+    @Override
+    public List<InventoryAsset> createInventoryAsset(List<Asset> assets) {
+        List<InventoryAsset> inventoryAssets = new ArrayList<>();
+        for (Asset asset : assets) {
+            inventoryAssets.add(new InventoryAsset(asset.getId(), asset.getDescription(),
+                    asset.getEmployeeInCharge(), asset.getValue(), asset.getDamage(), asset.getLocation(),
+                    Status.NOT_VERIFIED));
+        }
+        return inventoryAssets;
+    }
 }
